move get reportUrl into libraryLinks
